backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.11
pydantic==2.10.3
pydantic-settings==2.5.0
pydantic-core==2.27.1
python-dotenv==1.0.1
cors==1.0.1
google-auth==2.47.0
python-jose[cryptography]==3.3.0
resend==2.10.0
passlib[bcrypt]==1.7.4
